+++ lib/clearly/query/composer.rb
@@ -45,11 +45,31 @@ module Clearly
       # Composes a query from a parsed filter hash.
       # @param [ActiveRecord::Base] model
       # @param [Hash] hash
-      # @return [Arel::Nodes::Node, Array<Arel::Nodes::Node>]
+      # @return [ActiveRecord::Relation]
       def query(model, hash)
+        conditions = conditions(model, hash)
+        query = model.all
+        validate_query(query)
+        conditions.each do |condition|
+          validate_condition(condition)
+          query = query.where(condition)
+        end
+        query
+      end
+
+      # Composes Arel conditions from a parsed filter hash.
+      # @param [ActiveRecord::Base] model
+      # @param [Hash] hash
+      # @return [Array<Arel::Nodes::Node>]
+      def conditions(model, hash)
+        validate_model(model)
+        validate_hash(hash)
+
         definition = select_definition_from_model(model)
+        cleaned_query_hash = Clearly::Query::Cleaner.new.do(hash)
+
         # default combiner is :and
-        parse_query(definition, :and, hash)
+        parse_conditions(definition, :and, cleaned_query_hash)
       end
 
       private
@@ -82,7 +102,7 @@ module Clearly
       # @param [Symbol] query_key
       # @param [Hash] query_value
       # @return [Array<Arel::Nodes::Node>]
-      def parse_query(definition, query_key, query_value)
+      def parse_conditions(definition, query_key, query_value)
         if query_value.blank? || query_value.size < 1
           msg = "filter hash must have at least 1 entry, got '#{query_value.size}'"
           fail Clearly::Query::QueryArgumentError.new(msg, {hash: query_value})
@@ -130,8 +150,9 @@ module Clearly
       def parse_logical_operator(definition, logical_operator, value)
         validate_definition_instance(definition)
         validate_symbol(logical_operator)
+        validate_not_blank(value)
         validate_hash(value)
-        conditions = value.map { |key, value| parse_query(definition, key, value) }
+        conditions = value.map { |key, value| parse_conditions(definition, key, value) }
         condition_combine(logical_operator, *conditions)
       end
 
@@ -143,6 +164,7 @@ module Clearly
       def parse_standard_field(definition, field, value)
         validate_definition_instance(definition)
         validate_symbol(field)
+        validate_not_blank(value)
         validate_hash(value)
         value.map do |operator, operation_value|
           condition_components(operator, definition.table, field, definition.all_fields, operation_value)
@@ -159,6 +181,7 @@ module Clearly
         validate_symbol(field)
         fail Clearly::Query::QueryArgumentError.new('field name must contain a dot (.)') unless field.to_s.include?('.')
 
+        validate_not_blank(value)
         validate_hash(value)
 
         # extract table and field
@@ -189,6 +212,7 @@ module Clearly
         validate_definition_instance(definition)
         mapping = definition.get_field_mapping(field)
         validate_node_or_attribute(mapping)
+        validate_not_blank(value)
         validate_hash(value)
         value.map do |operator, operation_value|
           condition_node(operator, mapping, operation_value)
@@ -206,12 +230,12 @@ module Clearly
         [conditions].flatten.each { |c| validate_node_or_attribute(c) }
 
         current_model = definition.model
-        current_table = definition.table
+        #current_table = definition.table
         current_joins = definition.joins
 
         other_table = other_definition.table
         other_model = other_definition.model
-        other_joins = other_definition.joins
+        #other_joins = other_definition.joins
 
         # build an exist subquery to apply conditions that
         # refer to another table